Game tiles and code cleanup!
Made a new class to represent a game tile in a grid. The flow box now shows
these tiles, and the tile buttons will start the game associated with it when
pressed. Tiles have the ability to display an image/banner of their game on
their button. If no button is given it will just display a default message.
Also made flowbox expand on window resize instead of the old play button and
emulator selector.